Add files via upload
HTTP methods are used to define the action to be performed on a resource. Here are the methods you're using:

- app.get(): Retrieves data from the server. (Read operation)
- app.post(): Creates new data on the server. (Create operation)
- app.put(): Updates existing data on the server. (Update operation)
- app.delete(): Deletes data from the server. (Delete operation)

These HTTP methods are used to perform specific actions on users:
- POST /users: Creates a new user.
- PUT /users/:id: Updates an existing user with the specified ID.
- DELETE /users/:id: Deletes a user with the specified ID.
- GET /things/:name/:id: Retrieves data based on the provided name and ID.

Postman is used to send requests to these URLs; you're simulating HTTP requests to test your API endpoints.
The catch is handling any incoming requests that don't match any of the previously defined routes, returning a 404-status code with a message indicating that the URL is invalid.